+++ tanjun/__init__.py
@@ -170,6 +170,10 @@ __all__: list[str] = [
     "to_snowflake",
     "to_user",
     "to_voice_state",
+    # dependencies.py
+    "dependencies",
+    "LazyConstant",
+    "make_lc_resolver",
     # errors.py
     "errors",
     "CommandError",
@@ -205,8 +209,6 @@ __all__: list[str] = [
     "with_option",
     "with_multi_option",
     "with_parser",
-    # standard_dependencies.py
-    "standard_dependencies",
     # utilities.py
     "utilities",
 ]
@@ -214,7 +216,6 @@ __all__: list[str] = [
 import typing
 
 from . import abc
-from . import standard_dependencies
 from . import utilities
 from .checks import *
 from .clients import *
@@ -222,6 +223,7 @@ from .commands import *
 from .components import *
 from .context import *
 from .conversion import *
+from .dependencies import *
 from .errors import *
 from .hooks import *
 from .injecting import *